refactor: extract SpanCardContent into separate component
